+++ packages/monitor/dashboard/src/index.js
@@ -1,7 +1,4 @@
 import Vue from 'vue'
-import VueVirtualScroller from 'vue-virtual-scroller'
 import App from './components/App.vue'
-
-Vue.use(VueVirtualScroller)
 
 new Vue({ render: createElement => createElement(App) }).$mount('#app')