+++ packages/monitor/src/index.ts
@@ -1,4 +1,5 @@
 import {
+  defaultConfiguration,
   Extension,
   onChangePayload,
   onConfigurePayload,
@@ -60,7 +61,6 @@ export class Monitor implements Extension {
 
     if (this.configuration.enableDashboard) {
       this.dashboard = new Dashboard({
-        collector: this.collector,
         password: this.configuration.password,
         path: this.configuration.dashboardPath,
         port: this.configuration.port,
@@ -90,10 +90,12 @@ export class Monitor implements Extension {
     await this.storage.add('documentCount', await this.collector.documentCount())
     await this.storage.add('connectionCount', await this.collector.connectionCount())
     await this.storage.add('messageCount', await this.collector.messageCount())
+    await this.storage.set('documents', await this.collector.documents())
+    await this.storage.set('info', await this.collector.info())
   }
 
   private async cleanMetrics() {
-    const data = await this.storage.all()
+    const data = await this.storage.allTimed()
 
     for (let i = 0; i < data.length; i += 1) {
       if (moment(data[i].timestamp).add(1, 'hour').isBefore(moment())) {
@@ -140,21 +142,21 @@ export class Monitor implements Extension {
   }
 
   async onConnect(data: onConnectPayload) {
-    await this.storage.add('connections', this.collector.connect(data))
+    await this.storage.add('connectionLog', this.collector.connect(data))
   }
 
   async onDisconnect(data: onDisconnectPayload) {
-    await this.storage.add('connections', this.collector.disconnect(data))
+    await this.storage.add('connectionLog', this.collector.disconnect(data))
   }
 
   // eslint-disable-next-line @typescript-eslint/no-empty-function,no-empty-function
   async onCreateDocument(data: onCreateDocumentPayload) {
-    await this.storage.add('documents', this.collector.createDocument(data))
+    await this.storage.add('documentLog', this.collector.createDocument(data))
   }
 
   // eslint-disable-next-line @typescript-eslint/no-empty-function,no-empty-function
   async onChange(data: onChangePayload) {
-    await this.storage.add('documents', this.collector.changeDocument(data))
+    await this.storage.add('documentLog', this.collector.changeDocument(data))
   }
 
   // eslint-disable-next-line @typescript-eslint/no-empty-function,no-empty-function
@@ -170,9 +172,13 @@ export class Monitor implements Extension {
     this.collector.version = data.version
     this.collector.yjsVersion = data.yjsVersion
 
-    this.collector.serverConfiguration = {
-      port: data.configuration.port,
-      timeout: data.configuration.timeout,
-    }
+    const sanitizedConfiguration = {}
+
+    Object.keys(defaultConfiguration).forEach(key => {
+      // @ts-ignore
+      sanitizedConfiguration[key] = data.configuration[key] || null
+    })
+
+    this.collector.serverConfiguration = sanitizedConfiguration
   }
 }