+++ packages/server/src/Hocuspocus.ts
@@ -12,6 +12,8 @@ import packageJson from '../package.json'
 export const defaultConfiguration = {
   port: 80,
   timeout: 30000,
+  throttle: 5,
+  banTime: 5,
 }
 
 /**
@@ -34,6 +36,10 @@ export class Hocuspocus {
   }
 
   documents = new Map()
+
+  connectionsByIp: Map<String, Array<number>> = new Map()
+
+  bannedIps: Map<String, number> = new Map()
 
   httpServer?: HTTPServer
 
@@ -143,11 +149,18 @@ export class Hocuspocus {
    */
   handleConnection(incoming: WebSocket, request: IncomingMessage, context: any = null): void {
 
+    // get the remote ip address
+    const ip = <String> request.headers['x-real-ip']
+      || request.headers['x-forwarded-for']
+      || request.socket.remoteAddress || ''
+
+    // throttle the connection
+    if (this.throttle(ip)) {
+      return incoming.close()
+    }
+
     // create a unique identifier for every socket connection
     const socketId = uuid()
-
-    // @ts-ignore
-    incoming.socketId = socketId
 
     const hookPayload = {
       documentName: Hocuspocus.getDocumentName(request),
@@ -166,7 +179,7 @@ export class Hocuspocus {
       .then(() => {
         // if no hook interrupts create a document and connection
         const document = this.createDocument(request, socketId, context)
-        this.createConnection(incoming, request, document, context)
+        this.createConnection(incoming, request, document, socketId, context)
       })
       .catch(e => {
         // if a hook interrupts, close the websocket connection
@@ -180,7 +193,7 @@ export class Hocuspocus {
    * Handle update of the given document
    * @private
    */
-  private handleDocumentUpdate(document: Document, connection: Connection, update: Uint8Array, request: IncomingMessage, socketId: string): void {
+  private handleDocumentUpdate(document: Document, connection: Connection, update: Uint8Array, request: IncomingMessage, socketId: String): void {
 
     const hookPayload = {
       clientsCount: document.connectionsCount(),
@@ -214,7 +227,7 @@ export class Hocuspocus {
     const document = new Document(documentName)
 
     document.onUpdate((document, connection, update) => {
-      this.handleDocumentUpdate(document, connection, update, request, socketId)
+      this.handleDocumentUpdate(document, connection, update, request, connection.socketId)
     })
 
     const hookPayload = {
@@ -243,17 +256,16 @@ export class Hocuspocus {
    * Create a new connection by the given request and document
    * @private
    */
-  private createConnection(connection: WebSocket, request: IncomingMessage, document: Document, context?: any): Connection {
+  private createConnection(connection: WebSocket, request: IncomingMessage, document: Document, socketId: String, context?: any): Connection {
 
-    const instance = new Connection(connection, request, document, this.configuration.timeout, context)
+    const instance = new Connection(connection, request, document, this.configuration.timeout, socketId, context)
 
     instance.onClose(document => {
       const hookPayload = {
         clientsCount: document.connectionsCount(),
         context,
         document,
-        // @ts-ignore
-        socketId: connection.socketId,
+        socketId,
         documentName: document.name,
         requestHeaders: request.headers,
         requestParameters: Hocuspocus.getParameters(request),
@@ -313,6 +325,41 @@ export class Hocuspocus {
    */
   private static getDocumentName(request: IncomingMessage): string {
     return request.url?.slice(1)?.split('?')[0] || ''
+  }
+
+  /**
+   * Throttle requests
+   * @private
+   */
+  private throttle(ip: String): Boolean {
+    if (!this.configuration.throttle) {
+      return false
+    }
+
+    const bannedAt = this.bannedIps.get(ip) || 0
+
+    if (Date.now() < (bannedAt + (this.configuration.banTime * 60 * 1000))) {
+      return true
+    }
+
+    this.bannedIps.delete(ip)
+
+    // add this connection try to the list of previous connections
+    const previousConnections = this.connectionsByIp.get(ip) || []
+    previousConnections.push(Date.now())
+
+    // calculate the previous connections in the last minute
+    const previousConnectionsInTheLastMinute = previousConnections
+      .filter(timestamp => timestamp + (60 * 1000) > Date.now())
+
+    this.connectionsByIp.set(ip, previousConnectionsInTheLastMinute)
+
+    if (previousConnectionsInTheLastMinute.length > this.configuration.throttle) {
+      this.bannedIps.set(ip, Date.now())
+      return true
+    }
+
+    return false
   }
 }