+++ packages/monitor/CHANGELOG.md
@@ -3,6 +3,118 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0-alpha.33](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/monitor@1.0.0-alpha.32...@hocuspocus/monitor@1.0.0-alpha.33) (2021-03-10)
+
+**Note:** Version bump only for package @hocuspocus/monitor
+
+
+
+
+
+# [1.0.0-alpha.32](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/monitor@1.0.0-alpha.31...@hocuspocus/monitor@1.0.0-alpha.32) (2021-03-10)
+
+**Note:** Version bump only for package @hocuspocus/monitor
+
+
+
+
+
+# [1.0.0-alpha.31](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/monitor@1.0.0-alpha.30...@hocuspocus/monitor@1.0.0-alpha.31) (2021-03-10)
+
+**Note:** Version bump only for package @hocuspocus/monitor
+
+
+
+
+
+# [1.0.0-alpha.30](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/monitor@1.0.0-alpha.29...@hocuspocus/monitor@1.0.0-alpha.30) (2021-03-10)
+
+**Note:** Version bump only for package @hocuspocus/monitor
+
+
+
+
+
+# [1.0.0-alpha.29](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/monitor@1.0.0-alpha.28...@hocuspocus/monitor@1.0.0-alpha.29) (2021-03-10)
+
+**Note:** Version bump only for package @hocuspocus/monitor
+
+
+
+
+
+# [1.0.0-alpha.28](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/monitor@1.0.0-alpha.27...@hocuspocus/monitor@1.0.0-alpha.28) (2021-03-10)
+
+**Note:** Version bump only for package @hocuspocus/monitor
+
+
+
+
+
+# [1.0.0-alpha.27](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/monitor@1.0.0-alpha.26...@hocuspocus/monitor@1.0.0-alpha.27) (2021-03-10)
+
+**Note:** Version bump only for package @hocuspocus/monitor
+
+
+
+
+
+# [1.0.0-alpha.26](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/monitor@1.0.0-alpha.25...@hocuspocus/monitor@1.0.0-alpha.26) (2021-03-10)
+
+**Note:** Version bump only for package @hocuspocus/monitor
+
+
+
+
+
+# [1.0.0-alpha.25](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/monitor@1.0.0-alpha.24...@hocuspocus/monitor@1.0.0-alpha.25) (2021-03-10)
+
+**Note:** Version bump only for package @hocuspocus/monitor
+
+
+
+
+
+# [1.0.0-alpha.24](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/monitor@1.0.0-alpha.23...@hocuspocus/monitor@1.0.0-alpha.24) (2021-03-10)
+
+**Note:** Version bump only for package @hocuspocus/monitor
+
+
+
+
+
+# [1.0.0-alpha.23](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/monitor@1.0.0-alpha.22...@hocuspocus/monitor@1.0.0-alpha.23) (2021-03-10)
+
+**Note:** Version bump only for package @hocuspocus/monitor
+
+
+
+
+
+# [1.0.0-alpha.22](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/monitor@1.0.0-alpha.21...@hocuspocus/monitor@1.0.0-alpha.22) (2021-03-09)
+
+**Note:** Version bump only for package @hocuspocus/monitor
+
+
+
+
+
+# [1.0.0-alpha.21](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/monitor@1.0.0-alpha.20...@hocuspocus/monitor@1.0.0-alpha.21) (2021-03-09)
+
+**Note:** Version bump only for package @hocuspocus/monitor
+
+
+
+
+
+# [1.0.0-alpha.20](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/monitor@1.0.0-alpha.19...@hocuspocus/monitor@1.0.0-alpha.20) (2021-03-09)
+
+**Note:** Version bump only for package @hocuspocus/monitor
+
+
+
+
+
 # [1.0.0-alpha.19](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/monitor@1.0.0-alpha.18...@hocuspocus/monitor@1.0.0-alpha.19) (2021-03-04)
 
 **Note:** Version bump only for package @hocuspocus/monitor